+++ chart/spec/configuration/priorityclass_spec.rb
@@ -103,6 +103,75 @@ describe 'local priorityClass configuration' do
     ))
   end
 
+  let(:multiple_deployments_values) do
+    HelmTemplate.with_defaults(%(
+      global:
+        priorityClassName: system-cluster-critical
+      certmanager:
+        global:
+          priorityClassName: system-cluster-noncritical
+      certmanager-issuer:
+        priorityClassName: system-cluster-noncritical
+      gitlab:
+        geo-logcursor:
+          priorityClassName: system-cluster-noncritical
+        gitaly:
+          priorityClassName: system-cluster-noncritical
+        gitlab-exporter:
+          priorityClassName: system-cluster-noncritical
+        gitlab-pages:
+          priorityClassName: system-cluster-noncritical
+        gitlab-shell:
+          priorityClassName: system-cluster-noncritical
+        kas:
+          enabled: true  # DELETE THIS WHEN KAS BECOMES ENABLED BY DEFAULT
+          priorityClassName: system-cluster-noncritical
+        mailroom:
+          priorityClassName: system-cluster-noncritical
+        migrations:
+          priorityClassName: system-cluster-noncritical
+        praefect:
+          priorityClassName: system-cluster-noncritical
+        sidekiq:
+          pods:
+            - name: sidekiq-a
+              priorityClassName: system-cluster-noncritical
+            - name: sidekiq-b
+              priorityClassName: system-cluster-noncritical
+        spamcheck:
+          enabled: true  # DELETE THIS WHEN SPAMCHECK BECOMES ENABLED BY DEFAULT
+          priorityClassName: system-cluster-noncritical
+        toolbox:
+          priorityClassName: system-cluster-noncritical
+        webservice:
+          deployments:
+            webservice-a:
+              priorityClassName: system-cluster-noncritical
+              ingress:
+                path: /
+            webservice-b:
+              priorityClassName: system-cluster-noncritical
+              ingress:
+                path: /b
+      gitlab-runner:
+        priorityClassName: system-cluster-noncritical
+      minio:
+        priorityClassName: system-cluster-noncritical
+      nginx-ingress:
+        controller:
+          priorityClassName: system-cluster-noncritical
+      prometheus:
+        server:
+          priorityClassName: system-cluster-noncritical
+      registry:
+        priorityClassName: system-cluster-noncritical
+      shared-secrets:
+        priorityClassName: system-cluster-noncritical
+      upgradeCheck:
+        priorityClassName: system-cluster-noncritical
+    ))
+  end
+
   let(:ignored_deployments) do
     [
       'Deployment/test-certmanager',
@@ -120,6 +189,25 @@ describe 'local priorityClass configuration' do
   context 'When setting local priorityClassName' do
     it 'Populates priorityClassName for all deployments and jobs' do
       t = HelmTemplate.new(default_values)
+      expect(t.exit_code).to eq(0)
+
+      deployments = t.resources_by_kind('Deployment').reject { |key, _| ignored_deployments.include? key }
+
+      deployments.each do |key, _|
+        expect(t.dig(key, 'spec', 'template', 'spec', 'priorityClassName')).to eq('system-cluster-noncritical')
+      end
+
+      jobs = t.resources_by_kind('Job').reject { |key, _| ignored_jobs.include? key }
+
+      jobs.each do |key, _|
+        expect(t.dig(key, 'spec', 'template', 'spec', 'priorityClassName')).to eq('system-cluster-noncritical')
+      end
+    end
+  end
+
+  context 'When setting local per-deployment priorityClassName' do
+    it 'Populates priorityClassName for all deployments and jobs' do
+      t = HelmTemplate.new(multiple_deployments_values)
       expect(t.exit_code).to eq(0)
 
       deployments = t.resources_by_kind('Deployment').reject { |key, _| ignored_deployments.include? key }