+++ CHANGELOG.md
@@ -3,9 +3,26 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## [9.1.2-bb.0] (2025-07-15)
+## [9.2.1-bb.0] (2025-07-24)
 ### Changed
 - update gluon 0.6.2 -> 0.6.3
+- registry1.dso.mil/ironbank/gitlab/gitlab/certificates 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitaly 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-base 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-container-registry 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-exporter 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-mailroom 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-pages 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-shell 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-sidekiq 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-toolbox 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-webservice 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-workhorse 18.1.2 -> 18.2.1
+- registry1.dso.mil/ironbank/gitlab/gitlab/kubectl 18.1.2 -> 18.2.1
+
+## [9.1.2-bb.0] (2025-07-15)
+### Changed
+- update gluon 0.6.3 -> 0.7.0
 - registry1.dso.mil/ironbank/bitnami/redis 8.0.2 -> 8.0.3
 - registry1.dso.mil/ironbank/gitlab/gitlab/certificates 18.1.0 -> 18.1.2
 - registry1.dso.mil/ironbank/gitlab/gitlab/gitaly 18.1.0 -> 18.1.2